requirements_dev.txt
-r requirements.txt
ptvsd==4.2.0
pylint==2.2.2
autopep8==1.4.3
coverage==4.5.2
coveralls==1.5.1
pre-commit==1.12.0
tox
pytest
flake8-django
pylint-django